Tried out a less dynamic dynamic camera
The Camera is the player's sense of place and perspective. If it is
stable the world feels stable, if it is fluid the game feels fluid. If
the player notices it moving without telling it to move we lose a bit of
immersion, if it doesn't move when the player wants it to, we lose
immersion. I think this is a better balance but still probably too
stiff. We'll see